refactor: Move to commands directory.
