+++ src/scripts/commands/dev.ts
@@ -4,7 +4,7 @@ import config from '@shared/features/config';
 import express from 'express';
 import webpack from 'webpack';
 import WebpackShellPlugin from 'webpack-shell-plugin-next';
-import { webpackServerConfigure } from '../webpack.config';
+import { webpackServerConfigure } from '../../webpack.config';
 
 const scriptDev = async () => {
   const webpackServerConfig = webpackServerConfigure((webpackConfig) => {
